viewer: don't repeatedly show overlapping annotations
See GH issue #520.

This is problematic because it will break mouseover events on the
highlight's parent, but it does make the page not freeze when you mouse
in and out of a chunk of text with hundreds of annotations.  See
<http://willthompson.co.uk/misc/annotator-freeze-on-mouseout/> for an
example without this patch.

One alternative would be to use a single span.annotation-hl with
multiple annotation-ids where currently they would be nested. ie,
transform this:

```html
<span class="annotation-hl" data-annotation-id="123">
  <span class="annotator-hl" data-annotation-id="456">
    ...
      Foo
    ...
  </span>
</span>
```

into this:

```html
<span class="annotation-hl" data-annotation-id="123 456 ...">
  Foo
</span>
```

Another alternative would be to stick with the current structure but
make the .done() callback inside _onHighlightMouseover smarter about not
repeating itself.